+++ R/read_airports.R
@@ -6,8 +6,12 @@
 #' The data dictionary for public airports can be found at \url{https://www.anac.gov.br/acesso-a-informacao/dados-abertos/areas-de-atuacao/aerodromos/lista-de-aerodromos-publicos-v2/70-lista-de-aerodromos-publicos-v2}.
 #' The data dictionary for private airports can be found at \url{https://www.anac.gov.br/acesso-a-informacao/dados-abertos/areas-de-atuacao/aerodromos/lista-de-aerodromos-privados-v2}.
 #'
-#' @param type String. Whether function should download data  on `public` (Default)
-#'             or `private` airports.
+#' @param type String. Whether the function should download data on `all`,
+#'             `public` or `private` airports. Defaults to `all`, returning fewer
+#'             columns. Downloading `public` and `private` airports separately
+#'             will return the full set of columns available for each of those
+#'             data sets.
+#'
 #' @param showProgress Logical. Defaults to `TRUE` display progress.
 #' @return A `"data.table" "data.frame"` object.
 #' @export
@@ -15,22 +19,25 @@
 #' @importFrom data.table %like%
 #' @examples \dontrun{ if (interactive()) {
 #' # Read airports data
+#' all_airports <- read_airports(type = 'all')
+#'
 #' public_airports <- read_airports(type = 'public')
 #'
 #' private_airports <- read_airports(type = 'private')
 #'}}
-read_airports <- function(type = 'public', showProgress = TRUE){
+read_airports <- function(type = 'all', showProgress = TRUE){
 
 ### check inputs
-  if( ! type %in% c('public', 'private') ){ stop(paste0("Argument 'type' must be either 'public' or 'private'")) }
+  if( ! type %in% c('public', 'private', 'all') ){ stop(paste0("Argument 'type' must be either 'all, 'public' or 'private'")) }
   if( ! is.logical(showProgress) ){ stop(paste0("Argument 'showProgress' must be either 'TRUE' or 'FALSE.")) }
 
   # data url
+  # https://www.gov.br/anac/pt-br/assuntos/regulados/aerodromos/lista-de-aerodromos-civis-cadastrados
   url_public <- 'https://www.gov.br/anac/pt-br/assuntos/regulados/aerodromos/cadastro-de-aerodromos/aerodromos-cadastrados/cadastro-de-aerodromos-civis-publicos.csv'
   url_private <- 'https://sistemas.anac.gov.br/dadosabertos/Aerodromos/Lista%20de%20aer%C3%B3dromos%20privados/Aerodromos%20Privados/AerodromosPrivados.csv'
 
 ### download public airports
-if (type=='public'){
+if (any(type %in% c('public', 'all'))){
 
   dt_public <- try(silent=T,
                    data.table::fread(url_public,
@@ -50,11 +57,12 @@ if (type=='public'){
   dt_public <- latlon_to_numeric(df=dt_public, colname = 'LATITUDE')
   dt_public <- latlon_to_numeric(df=dt_public, colname = 'LONGITUDE')
 
-  return(dt_public)
+  # names to to lower case
+  data.table::setnames(dt_public, tolower(names(dt_public)))
   }
 
 ### download private airports
-else if (type=='private'){
+if (any(type %in% c('private', 'all'))){
 
   dt_private <- try(silent=T,
                     data.table::fread(url_private,
@@ -69,10 +77,26 @@ else if (type=='private'){
   dt_private <- latlon_to_numeric(df=dt_private, colname = 'Latitude')
   dt_private <- latlon_to_numeric(df=dt_private, colname = 'Longitude')
 
-  ## last time the data was updated
-  # last_update_private <- data.table::fread(url_private, nrows = 1)
-  # last_update_private[1,]
-  return(dt_private)
+  # names to to lower case
+  data.table::setnames(dt_private, tolower(names(dt_private)))
   }
+
+if (type == 'private') { return(dt_private) }
+if (type == 'public') { return(dt_public) }
+if (type == 'all') {
+                    ## find columns in common
+                    # cols_to_keep <- c("código oaci", "ciad", "nome", "uf", "longitude", "latitude", "altitude")
+                    a <- names(dt_private)
+                    b <- names(dt_public)
+                    cols_to_keep <- a[a %in% b]
+
+                    ## subset columns in common
+                    dt_public <- dt_public[, cols_to_keep, with=FALSE]
+                    dt_private <- dt_private[, cols_to_keep, with=FALSE]
+
+                    dt <- data.table::rbindlist(list(dt_public, dt_private), use.names=FALSE)
+
+                    return(dt)
+                    }
 }