+++ NEWS.md
@@ -5,7 +5,7 @@
 # v0.1.0
 
 * Major changes:
-- New function `read_airports()` to read data on all public and private airports. [Closes #4](https://github.com/ipeaGIT/flightsbr/issues/4).
+- New function `read_airports()` to read data on all public and private airports. [Closes #4](https://github.com/ipeaGIT/flightsbr/issues/4) and [Closes #9](https://github.com/ipeaGIT/flightsbr/issues/9).
 - Function `read_flighs()` now takes `date` input in the format `yyyymm` or `yyyy`. The the date is a 4-digit number, the function now downloads the data of all months in that year. [Closes #2](https://github.com/ipeaGIT/flightsbr/issues/1).
 - new internal support functions:
   - `split_date()`: Split a date from yyyymmm to year yyyy and month mm